Move init sidebar functionality to sidebar base
